Move mockLogger to its own file
