+++ packages/server/src/__tests__/mockLogger.ts
@@ -1,0 +1,8 @@
+export function mockLogger() {
+  return {
+    debug: jest.fn(),
+    info: jest.fn(),
+    warn: jest.fn(),
+    error: jest.fn(),
+  };
+}
